moving steps from SharedSteps to ParsingSteps
